chore(Report): Bump Viz Versions
Sync report with ECL Watch DataPatterns Report
