+++ .github/workflows/build-test-report.yml
@@ -1,0 +1,34 @@
+name: Build Test ECL Watch
+
+on:
+  push:
+    branches:
+      - "master"
+    paths:
+      - "report/**/*"
+  pull_request:
+    branches:
+      - "master"
+    paths:
+      - "report/**/*"
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v2
+        with:
+          submodules: "recursive"
+      - uses: actions/setup-node@v1
+        with:
+          node-version: "12.x"
+      - name: Install Dependencies
+        working-directory: ./report
+        run: npm ci
+      - name: Build
+        working-directory: ./report
+        run: npm run build
+      - name: Test
+        working-directory: ./report
+        run: npm run test
