+++ report/package.json
@@ -16,22 +16,23 @@
     "watch": "run-p compile-es6-watch bundle-watch",
     "link-hpcc-js": "mklink /J hpcc-js ..\\hpcc-js",
     "lint": "tslint --project . src/**/*.ts",
-    "docs": "typedoc --options tdoptions.json ."
+    "docs": "typedoc --options tdoptions.json .",
+    "test": "run-s lint"
   },
   "dependencies": {
-    "@hpcc-js/loader": "2.6.4",
-    "tslib": "1.9.3"
+    "@hpcc-js/loader": "2.13.2",
+    "tslib": "1.11.1"
   },
   "devDependencies": {
-    "npm-run-all": "4.1.5",
-    "rimraf": "2.6.3",
-    "rollup": "1.10.0",
-    "rollup-plugin-alias": "1.5.1",
-    "rollup-plugin-commonjs": "9.3.4",
-    "rollup-plugin-node-resolve": "4.2.3",
-    "rollup-plugin-postcss": "2.0.3",
-    "terser": "3.17.0",
-    "typescript": "3.4.3"
+    "npm-run-all": "^4.1.5",
+    "rimraf": "^3.0.2",
+    "rollup": "^1.32.1",
+    "rollup-plugin-alias": "^1.5.2",
+    "rollup-plugin-commonjs": "^9.3.4",
+    "rollup-plugin-node-resolve": "^4.2.4",
+    "rollup-plugin-postcss": "^2.5.0",
+    "terser": "^4.6.7",
+    "typescript": "^3.8.3"
   },
   "repository": {
     "type": "git",
