+++ report/src/config.ts
@@ -1,0 +1,18 @@
+export const config = {
+    rowHeight: 190,
+    colRatios: {
+        attributeDesc: 1,
+        statsData: 1,
+        breakdown: 2 / 3,
+        quartile: 1,
+        popularPatterns: 1
+    },
+    primaryFontSize: 20,
+    secondaryFontSize: 13,
+    primaryColor: "#494949",
+    secondaryColor: "#DDD",
+    offwhiteColor: "#FBFBFB",
+    blueColor: "#1A99D5",
+    redColor: "#ED1C24",
+    fillRateRedThreshold: 33
+};
