+++ report/tsconfig.json
@@ -16,7 +16,8 @@
         "lib": [
             "dom",
             "es5",
-            "es2015.promise"
+            "es2015.promise",
+            "es2015.Iterable"
         ],
         "types": []
     },
